add exception handling for product update and delete operations

+++ src/main/java/com/mayan/ecommerce/services/ProductService.java
@@ -3,11 +3,15 @@ package com.mayan.ecommerce.services;
 import com.mayan.ecommerce.dtos.ProductDTO;
 import com.mayan.ecommerce.entities.Product;
 import com.mayan.ecommerce.repositories.ProductRepository;
+import com.mayan.ecommerce.services.exceptions.DatabaseException;
 import com.mayan.ecommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 
@@ -41,15 +45,25 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto) {
+        try {
         Product result = repository.getReferenceById(id);
         copyDtoToEntity(dto, result);
         result = repository.save(result);
         return new ProductDTO(result);
+        } catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException("Produto não encontrado.");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Produto não encontrado.");
+        } try {
         repository.deleteById(id);
+        } catch (DataIntegrityViolationException e){
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
 
     private void copyDtoToEntity(ProductDTO dto, Product entity) {
